# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Book of Memory API server.",
        "title": "Book of Memory API",
        "contact": {
            "name": "API Support",
            "url": "https://t.me/azat_bil",
            "email": "az@bilalov@mail.ru"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/bindings": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Find bindings",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binding"
                ],
                "summary": "Find bindings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Date from",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Date to",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.Binding"
                            }
                        }
                    },
                    "400": {
                        "description": "некорректный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "отсутствует токен авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "токен авторизации в блеклисте",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bindings/{uuid}": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Find binding by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binding"
                ],
                "summary": "Find binding by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Binding"
                        }
                    },
                    "400": {
                        "description": "идентификатор пользователя пустой или недействительный",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "отсутствует токен авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "токен авторизации в блеклисте",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Update binding by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binding"
                ],
                "summary": "Update binding by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.BindingUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Binding"
                        }
                    },
                    "400": {
                        "description": "статус не является entered или in_progress",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "отсутствует токен авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "токен авторизации в блеклисте",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Delete binding by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binding"
                ],
                "summary": "Delete binding by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "статус не является entered",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "отсутствует токен авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "токен авторизации в блеклисте",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bindings/{uuid}/accept-reject": {
            "put": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Accept or reject binding by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binding"
                ],
                "summary": "Accept or reject binding by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.BindingStatusUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Binding"
                        }
                    },
                    "400": {
                        "description": "статус недействителен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "отсутствует токен авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "токен авторизации в блеклисте",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bindings/{uuid}/submit": {
            "put": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Submit binding by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binding"
                ],
                "summary": "Submit binding by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Binding"
                        }
                    },
                    "400": {
                        "description": "ветеран не установлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "отсутствует токен авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "токен авторизации в блеклисте",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Binding": {
            "type": "object",
            "properties": {
                "binding_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Document"
                    }
                },
                "endedAt": {
                    "type": "string"
                },
                "formattedAt": {
                    "type": "string"
                },
                "moderator": {
                    "$ref": "#/definitions/ds.User"
                },
                "moderatorID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/ds.User"
                },
                "userID": {
                    "type": "string"
                },
                "veteran": {
                    "$ref": "#/definitions/ds.Veteran"
                },
                "veteranID": {
                    "type": "string"
                }
            }
        },
        "ds.BindingStatusUpdateRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "ds.BindingUpdateRequest": {
            "type": "object",
            "properties": {
                "veteran_id": {
                    "type": "string"
                }
            }
        },
...
        "ds.Document": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "ds.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "passwd": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "ds.Veteran": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "veteran_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}